requirements.txt
aiogram==3.*
aiosqlite
python-dotenv